consolidate packages and sdks service to registry service
