+++ api-node/src/app.module.ts
@@ -1,13 +1,11 @@
 import { Module } from '@nestjs/common';
 import { PackagesController } from './packages/packages.controller';
-import { PackagesService } from './packages/packages.service';
 import { HealthCheckController } from './health/healthCheck.controller';
 import { MarketingController } from './marketing/marketing.controller';
 import { MarketingService } from './marketing/marketing.service';
 import { AppsController } from './apps/apps.controller';
 import { AppsService } from './apps/apps.service';
 import { SdksController } from './sdks/sdks.controller';
-import { SdksService } from './sdks/sdks.service';
 import { AwsLambdaLayersController } from './aws-lambda-layers/aws-lambda-layers.controller';
 import { RegistryService } from './registry/registry.service';
 
@@ -21,6 +19,6 @@ import { RegistryService } from './registry/registry.service';
     SdksController,
     AwsLambdaLayersController,
   ],
-  providers: [PackagesService, MarketingService, AppsService, SdksService, RegistryService],
+  providers: [MarketingService, AppsService, RegistryService],
 })
 export class AppModule {}
